Organize payment code
