+++ src/Admin/Models/BillingInformationModel.cs
@@ -1,4 +1,5 @@
 ﻿using Bit.Core.Billing.Models;
+using Bit.Core.Billing.Payment.Models;
 
 namespace Bit.Admin.Models;
