+++ src/Api/Billing/Models/Requests/TokenizedPaymentSourceRequestBody.cs
@@ -1,6 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using Bit.Api.Utilities;
 using Bit.Core.Billing.Models;
+using Bit.Core.Billing.Payment.Models;
 using Bit.Core.Enums;
 
 namespace Bit.Api.Billing.Models.Requests;
