+++ src/Api/Billing/Models/Responses/BillingPaymentResponseModel.cs
@@ -1,4 +1,5 @@
 ﻿using Bit.Core.Billing.Models;
+using Bit.Core.Billing.Payment.Models;
 using Bit.Core.Models.Api;
 
 namespace Bit.Api.Billing.Models.Responses;
