+++ src/Core/Billing/Organizations/Services/IOrganizationBillingService.cs
@@ -1,6 +1,7 @@
 ﻿using Bit.Core.AdminConsole.Entities;
 using Bit.Core.Billing.Models;
 using Bit.Core.Billing.Organizations.Models;
+using Bit.Core.Billing.Payment.Models;
 
 namespace Bit.Core.Billing.Organizations.Services;
