+++ src/Core/Billing/Payment/Models/BillingInfo.cs
@@ -1,7 +1,7 @@
 ﻿using Bit.Core.Enums;
 using Stripe;
 
-namespace Bit.Core.Billing.Models;
+namespace Bit.Core.Billing.Payment.Models;
 
 public class BillingInfo
 {
