+++ src/Core/Billing/Payment/Models/TokenizedPaymentSource.cs
@@ -1,6 +1,6 @@
 ﻿using Bit.Core.Enums;
 
-namespace Bit.Core.Billing.Models;
+namespace Bit.Core.Billing.Payment.Models;
 
 public record TokenizedPaymentSource(
     PaymentMethodType Type,
