+++ src/Core/Billing/Payment/Services/PaymentHistoryService.cs
@@ -1,13 +1,13 @@
 ﻿#nullable enable
 using Bit.Core.AdminConsole.Entities;
-using Bit.Core.Billing.Models;
+using Bit.Core.Billing.Payment.Models;
 using Bit.Core.Entities;
 using Bit.Core.Models.BitStripe;
 using Bit.Core.Repositories;
 using Bit.Core.Services;
 using Microsoft.Extensions.Logging;
 
-namespace Bit.Core.Billing.Services.Implementations;
+namespace Bit.Core.Billing.Payment.Services;
 
 public class PaymentHistoryService(
     IStripeAdapter stripeAdapter,
